webapp/requirements.txt
# WebApp Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.2
httpx>=0.25.0
pydantic[email]>=2.4.0
supabase>=2.0.0
asyncio>=3.4.3
typing-extensions>=4.8.0